Add and fix translations
