+++ src/pagePartials/creator/register/RegistrationSteps.tsx
@@ -154,7 +154,7 @@ export default function RegistrationSteps({ onSubmit, txState }: RegistrationSte
         </Typography>
 
         <Typography textAlign="justify" variant="body3" width="85%">
-          {t(`creator.register.steps.${currentStep}.sub-title`)}
+          {t(`creator.register.steps.${currentStep}.subTitle`)}
         </Typography>
       </Stack>
       <FormWithSteps
@@ -163,7 +163,7 @@ export default function RegistrationSteps({ onSubmit, txState }: RegistrationSte
           {},
           {
             terms: {
-              agreementText: t('creator.register.form.terms-conditions'),
+              agreementText: t('creator.register.form.termsConditions'),
             },
           },
         ]}
