+++ src/pagePartials/home/BadgesInReviewSwiper.tsx
@@ -1,22 +1,22 @@
-import { getPendingTimeProgressPercentage, getTimeLeft, TimeLeft, timestampToDate } from "@/src/utils/date";
 import { useRouter } from 'next/navigation'
 import { useMemo } from 'react'
 
 import { Box } from '@mui/material'
-import { EmptyBadgePreview, PendingBadgeOverlay } from "thebadge-ui-library";
+import { EmptyBadgePreview, PendingBadgeOverlay } from 'thebadge-ui-library'
 
 import InViewPort from '@/src/components/helpers/InViewPort'
 import SafeSuspense from '@/src/components/helpers/SafeSuspense'
 import TBSwiper from '@/src/components/helpers/TBSwiper'
 import { fillListWithPlaceholders } from '@/src/components/utils/emptyBadges'
 import useSubgraph from '@/src/hooks/subgraph/useSubgraph'
+import { TimeLeft, useDate } from '@/src/hooks/useDate'
 import BadgeTypeMetadata from '@/src/pagePartials/badge/BadgeTypeMetadata'
 
 const now = Math.floor(Date.now() / 1000)
 export default function BadgesInReviewSwiper() {
   const router = useRouter()
   const gql = useSubgraph()
-
+  const { getPendingTimeProgressPercentage, getTimeLeft, timestampToDate } = useDate()
   const badgesInReview = gql.useBadgesInReview({ date: now })
 
   const badgesList = useMemo(() => {
@@ -53,7 +53,13 @@ export default function BadgesInReviewSwiper() {
     })
     // If there is no badges to show, we list 5 placeholders
     return fillListWithPlaceholders(badges, <EmptyBadgePreview size="small" />, 5)
-  }, [badgesInReview.data?.badges, router])
+  }, [
+    badgesInReview.data?.badges,
+    getPendingTimeProgressPercentage,
+    getTimeLeft,
+    router,
+    timestampToDate,
+  ])
 
   return <TBSwiper items={badgesList} />
 }
