+++ src/pagePartials/profile/myProfile/PendingList.tsx
@@ -8,14 +8,14 @@ import SafeSuspense from '@/src/components/helpers/SafeSuspense'
 import TBSwiper from '@/src/components/helpers/TBSwiper'
 import { fillListWithPlaceholders } from '@/src/components/utils/emptyBadges'
 import useSubgraph from '@/src/hooks/subgraph/useSubgraph'
+import { TimeLeft, useDate } from '@/src/hooks/useDate'
 import BadgeTypeMetadata from '@/src/pagePartials/badge/BadgeTypeMetadata'
-import { TimeLeft, getPendingTimeProgressPercentage, getTimeLeft, timestampToDate } from '@/src/utils/date'
 
 const now = Math.floor(Date.now() / 1000)
 export default function PendingList() {
   const router = useRouter()
   const gql = useSubgraph()
-
+  const { getPendingTimeProgressPercentage, getTimeLeft, timestampToDate } = useDate()
   const badgesInReview = gql.useBadgesInReview({ date: now })
 
   const badgesList = useMemo(() => {
@@ -51,7 +51,13 @@ export default function PendingList() {
     })
     // If there is no badges to show, we list 5 placeholders
     return fillListWithPlaceholders(badges, <EmptyBadgePreview size="small" />, 5)
-  }, [badgesInReview.data?.badges, router])
+  }, [
+    badgesInReview.data?.badges,
+    getPendingTimeProgressPercentage,
+    getTimeLeft,
+    router,
+    timestampToDate,
+  ])
 
   return (
     <TBSwiper
